docs: rewrite project documentation and fix license plate bidi rendering
  - Replace empty README.md with full project overview (stack, quickstart,
    env vars, feature matrix)
  - Rewrite backend/README.md: Spring Boot 3.3.7, JWT/Security, Flyway V1-V12,
    Playwright PDF, Resend emails, complete endpoint reference, CMP rules
  - Rewrite frontend/README.md: all 17 routes, every component documented,
    accurate stack (Angular 17, no Bootstrap), Vercel deployment
  - Rewrite doc/API_DOCUMENTATION.md: all current endpoints with correct JSON
    examples (auth, bills, sales/combined, invoices/preview…)
  - Remove 11 stale ad-hoc reports at project root (EMAIL_ARCHITECTURE.md,
    PDF_UPGRADE_SUMMARY.md, RESEND_SETUP.md, etc.)
  - Fix Tunisian license plate bidi rendering in facture_v3/v4/v5 templates:
    replace single-string span with inline-flex three-part layout
    (licensePlateY / تونس / licensePlateX) to bypass Unicode bidi reordering

+++ backend/src/main/java/com/example/stock_management/dto/ProductDashboardDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -13,14 +15,14 @@ public class ProductDashboardDTO {
     private String name;
     private String category;
     private String unit;
-    private Double unitPrice;
-    private Double unitPriceSold;
+    private BigDecimal unitPrice;
+    private BigDecimal unitPriceSold;
     private Integer currentStockQuantity;
     private Long stockVendu;
     private Long stockEntrepot;
     private Long purchasesCount;
-    private Double averagePurchasePrice;
+    private BigDecimal averagePurchasePrice;
     private Long salesCount;
-    private Double averageSalePrice;
-    private Double bilan;
+    private BigDecimal averageSalePrice;
+    private BigDecimal bilan;
 }
