+++ backend/src/main/java/com/example/stock_management/dto/StockSummaryDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,10 +13,10 @@ public class StockSummaryDTO {
     private Long productId;
     private String productDesignation;
     private Integer initialQuantity;
-    private Double initialValue;
-    private Double totalPurchasesAmount; // Total achats en TTC
-    private Double totalSalesAmount; // Total ventes en TTC
-    private Integer finalQuantity; // stock final
-    private Double finalStockValue; // Valeur du stock final
-    private Double cmp; // Coût Moyen Pondéré
+    private BigDecimal initialValue;
+    private BigDecimal totalPurchasesAmount;
+    private BigDecimal totalSalesAmount;
+    private Integer finalQuantity;
+    private BigDecimal finalStockValue;
+    private BigDecimal cmp;
 }
